Fixing reflection bug

+++ Source/Tracer-Test/PathTracer.cu
@@ -160,7 +160,7 @@ bool PathTracer::Render()
     globalData.lightList = dLights;
     globalData.totalLightCount = lightCount;
     globalData.mediumList = dMediums;
-    globalData.totalMediumCount = scene.MediumsCPU().size();
+    globalData.totalMediumCount = static_cast<uint32_t>(scene.MediumsCPU().size());
 
     globalData.nee = options.nextEventEstimation;
     globalData.rrStart = options.rrStart;
